requirements.txt
aiofiles==24.1.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comm==0.2.3
ddgs==9.12.0
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.135.2
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
gradio==6.10.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
openai==2.30.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
primp==1.2.1
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stack-data==0.6.3
starlette==0.52.1
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
watchdog==6.0.0
wcwidth==0.6.0
